fect: 기능분리 1차
