+++ common/core/abstract/abstract_trade_api.py
@@ -1,12 +1,18 @@
 """대한민국 거래소 그리고 해외거래소 통합 추상 클래스"""
 
+from typing import Callable
 from typing import Coroutine, Any
 from abc import abstractmethod, ABC
 from common.core.types import ExchangeResponseData
+from common.utils.other_utils import get_symbol_collect_url
+from common.core.types import SubScribeFormat
 
 
 # Rest
 class AbstractExchangeRestClient(ABC):
+    def __init__(self, market: str) -> None:
+        self._rest = get_symbol_collect_url(market=market, type_="rest")
+
     @abstractmethod
     async def get_coin_all_info_price(self, coin_name: str) -> ExchangeResponseData:
         """
@@ -22,6 +28,14 @@ class AbstractExchangeRestClient(ABC):
 
 # Socket
 class AbstractExchangeSocketClient(ABC):
+    def __init__(
+        self, target: str, socket_parameter: Callable[[str], SubScribeFormat]
+    ) -> None:
+        self._websocket = get_symbol_collect_url(target, "socket")
+        self.socket_parameter = socket_parameter
+        self.ticker = "ticker"
+        self.orderbook = "orderbook"
+
     @abstractmethod
     async def get_present_websocket(
         self, symbol: str, req_type: str | None = None
