+++ foreign/client/rest_foreign_exchange.py
