+++ foreign/client/socket_foreign_exchange.py
@@ -2,10 +2,7 @@
 코인 정보 추상화
 """
 
-from typing import Callable
 from common.core.abstract import AbstractExchangeSocketClient
-from common.utils.other_utils import get_symbol_collect_url
-from common.core.types import SubScribeFormat
 from common.setting.socket_parameter import (
     binance_socket_paramater,
     kraken_socket_parameter,
@@ -16,24 +13,19 @@ from common.setting.socket_parameter import (
 
 
 class CoinExchangeSocketClient(AbstractExchangeSocketClient):
-    def __init__(
-        self, target: str, socket_parameter: Callable[[str], SubScribeFormat]
+    async def get_present_websocket(
+        self, symbol: str, req_type: str, socket_type: str
     ) -> None:
-        self._websocket = get_symbol_collect_url(target, "socket")
-        self.socket_parameter = socket_parameter
-        self.ticker = "ticker"
-        self.orderbook = "orderbook"
-
-    # fmt: off
-    async def get_present_websocket(self, symbol: str, req_type: str, socket_type: str) -> None:
-        from pipe.foreign.foreign_websocket_client import ForeignWebsocketConnection as WCM
+        from foreign.connection.foreign_websocket_connect import (
+            ForeignWebsocketConnection as WCM,
+        )
 
         """소켓 출발점"""
         return await WCM().websocket_to_json(
             uri=self._websocket,
             subs_fmt=self.socket_parameter(symbol=symbol, req_type=req_type),
             symbol=symbol,
-            socket_type=socket_type
+            socket_type=socket_type,
         )
